Fix delete post button | Change structure of files
