+++ src/app/_components/posts-wrapper/posts/components/PostDescription.tsx
@@ -1,0 +1,6 @@
+
+export function PostDescription({ postDescription }: { postDescription: string; }) {
+  return (
+    <p className="h-[100px] px-2 overflow-y-auto">{postDescription}</p>
+  );
+}
