+++ src/app/_components/posts-wrapper/posts/components/PostTitle.tsx
@@ -1,0 +1,6 @@
+
+export function PostTitle({ postTitle }: { postTitle: string; }) {
+  return (
+    <h2 className="card-title px-2">{postTitle}</h2>
+  );
+}
